flask-backend/requirements.txt
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
mysql-connector-python==8.0.24
numpy==1.20.3
Pillow==8.2.0
protobuf==3.16.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
Werkzeug==1.0.1
